reorganize runner
